+++ docs/bu-blocks-frontend.js
@@ -74,6 +74,38 @@ if (!Array.prototype.findIndex) {
   });
 }
 
+
+// Matches polyfill.
+if (!Element.prototype.matches) {
+	Element.prototype.matches =
+	  Element.prototype.msMatchesSelector ||
+	  Element.prototype.webkitMatchesSelector;
+}
+// element.closest() polyfill.
+if (!Element.prototype.closest) {
+	Element.prototype.closest = function(s) {
+	  var el = this;
+
+	  do {
+		if (Element.prototype.matches.call(el, s)) return el;
+		el = el.parentElement || el.parentNode;
+	  } while (el !== null && el.nodeType === 1);
+	  return null;
+	};
+}
+
+// Foreach NodeList Polyfill for IE.
+if ('NodeList' in window && !NodeList.prototype.forEach) {
+    console.info('polyfill for IE11');
+    NodeList.prototype.forEach = function (callback, thisArg) {
+      thisArg = thisArg || window;
+      for (var i = 0; i < this.length; i++) {
+        callback.call(thisArg, this[i], i, this);
+      }
+    };
+}
+
+
 // Foreach Polyfill
 // Production steps of ECMA-262, Edition 5, 15.4.4.18
 // Reference: http://es5.github.io/#x15.4.4.18
@@ -510,7 +542,385 @@ const bu_blocks = {};
 			setButtonText( str );
 		}
 	};
-})();;bu_blocks.drawer = ( function() {
+})();;bu_blocks.collapsibleControl = ( function() {
+
+	// Store all Control blocks
+	var collapsibleControlBlocks = [];
+	var allCollapsibleBlocks = [];
+	var allBlocksOpen = false;
+	var collapsibleOpenClass = 'is-open';
+
+	/**
+	 * Open or close a group of collapsible blocks
+	 *
+	 * @param array collapsible blocks
+	 * @param bool true to open set of collapsible blocks, false to close
+	 */
+	var controlCollapsibleBlocks = function( collapsibleBlocks, open ) {
+
+		if ( open === undefined ) {
+			open = true;
+		}
+
+		collapsibleBlocks.forEach( function( collapsible, i ) {
+			const container = collapsible.container;
+			const toggle = collapsible.toggle;
+			const panel = collapsible.panel;
+
+			if ( open ) {
+				container.classList.add( collapsibleOpenClass );
+				toggle.setAttribute( 'aria-expanded', true );
+				panel.setAttribute( 'aria-hidden', false );
+			} else {
+				container.classList.remove( collapsibleOpenClass );
+				toggle.setAttribute( 'aria-expanded', false );
+				panel.setAttribute( 'aria-hidden', true );
+			}
+
+		} );
+
+	}
+
+	/**
+	 * Toggle all Collapsible blocks
+	 */
+	var toggleAll = function( control ) {
+
+		if ( 0 === allCollapsibleBlocks.length ) {
+			return;
+		}
+
+		controlCollapsibleBlocks( allCollapsibleBlocks, !allBlocksOpen );
+		allBlocksOpen = ( allBlocksOpen ) ? false : true;
+	};
+
+	/**
+	 * Toggle Collapsible blocks in control's group
+	 */
+	var toggleGroup = function( control ) {
+		const groupIsOpen = control.groupIsOpen;
+		const collapsibleBlocks = control.collapsibleBlocks
+
+		controlCollapsibleBlocks( collapsibleBlocks, !groupIsOpen );
+		control.groupIsOpen = ( groupIsOpen ) ? false : true;
+	};
+
+	/**
+	 * Find all Collapsible blocks on a page
+	 */
+	var findAllCollapsibleBlocks = function() {
+		var containers = document.querySelectorAll( '.js-wp-block-bu-collapsible' );
+
+		// Don't coninue if no Collapsible blocks exist
+		if ( containers.length === 0 ) {
+			return;
+		}
+
+		containers.forEach( function( element, i ) {
+			var block = {};
+
+			block.container = element;
+			block.toggle = element.querySelector( '.js-bu-block-collapsible-toggle' );
+			block.panel = element.querySelector( '.js-bu-block-collapsible-content' );
+			allCollapsibleBlocks.push( block );
+		} );
+
+	};
+
+	/**
+	 * Return all Collapsible blocks in the group with a Control
+	 *
+	 * @param object control
+	 *
+	 * @return array list of all collapsible blocks in group
+	 */
+	var getGroupCollapsibleBlocks = function( control ) {
+
+		var blocks = [];
+		var group = control.closest( '.wp-block-group' );
+		if ( ! group ) {
+			return blocks;
+		}
+		var containers = group.querySelectorAll( '.js-wp-block-bu-collapsible' );
+
+		containers.forEach( function( element, i ) {
+			var block = {};
+
+			block.container = element;
+			block.toggle = element.querySelector( '.js-bu-block-collapsible-toggle' );
+			block.panel = element.querySelector( '.js-bu-block-collapsible-content' );
+			blocks.push( block );
+		} );
+
+		return blocks;
+	}
+
+	/**
+	 * Find all Controls and Collapsible blocks
+	 */
+	var findElements = function() {
+
+		var controls = document.querySelectorAll( '.bu-collapsible-control-toggle' );
+		var allCollapsibleBlocksFound = false;
+
+		// Don't coninue if no Controls are found
+		if ( controls.length === 0 ) {
+			return;
+		}
+
+		// Store all controls
+		controls.forEach( function( control, i ) {
+			var block = {};
+
+			block.toggle = control;
+
+			// Check if Control targets all blocks or blocks in its group
+			if ( control.classList.contains( 'js-bu-collapsible-control-target-group' ) ) {
+				block.targetGroup = true;
+				block.collapsibleBlocks = getGroupCollapsibleBlocks( control );
+				block.groupIsOpen = false;
+			} else {
+				block.targetGroup = false;
+
+				if ( ! allCollapsibleBlocksFound ) {
+					findAllCollapsibleBlocks();
+				}
+
+				allCollapsibleBlocksFound = true;
+			}
+
+			collapsibleControlBlocks.push( block );
+		} );
+
+	};
+
+	/**
+	 * Set up handlers, aria, and other functionality
+	 */
+	var setupCollapsibleControlBlocks = function() {
+		if ( collapsibleControlBlocks.length === 0 ) {
+			return;
+		}
+
+		collapsibleControlBlocks.forEach( function( control, i ) {
+			const toggle = control.toggle;
+			const targetGroup = control.targetGroup;
+
+			toggle.addEventListener( 'click', function( e ) {
+				e.preventDefault();
+				if ( targetGroup ) {
+					toggleGroup( control );
+				} else {
+					toggleAll( control );
+				}
+			} );
+		} );
+
+	};
+
+	/**
+	 * Init
+	 */
+	var collapsibleControlInit = function() {
+		findElements();
+		setupCollapsibleControlBlocks();
+	};
+
+	// Start things on dom ready.
+	document.addEventListener( "DOMContentLoaded", function() {
+		collapsibleControlInit();
+	} );
+
+} )();
+;bu_blocks.collapsible = ( function() {
+
+	// Store all collapsible block
+	var collapsibleBlocks = [];
+	var collapsibleOpenClass = 'is-open';
+	var collapsibleClosedClass = 'is-closed';
+	var eventOpen = new CustomEvent('bu-blocks-collapsible-open');
+	var eventClose = new CustomEvent('bu-blocks-collapsible-close');
+
+
+
+
+	/**
+	 * Check if a Collapsible block is set to open by default by user.
+	 *
+	 * @param object collapsible block
+	 * @return bool
+	 */
+	var isOpenDefault = function( collapsible ) {
+		const container = collapsible.container;
+
+		if ( 'true' === container.getAttribute("data-default-open") ) {
+			return true;
+		}
+
+		return false;
+	};
+
+	/**
+	 * Check if a Collapsible block is open.
+	 *
+	 * @param object collapsible block
+	 * @return bool
+	 */
+	var isOpen = function( collapsible ) {
+		const container = collapsible.container;
+
+		if ( container.classList.contains ( collapsibleOpenClass ) ) {
+			return true;
+		}
+
+		return false;
+	};
+
+	/**
+	 * Open Collapsible block
+	 *
+	 * @param object collapsible block
+	 */
+	var openCollapsible = function( collapsible ) {
+		const container = collapsible.container;
+		const toggle = collapsible.toggle;
+		const panel = collapsible.panel;
+
+		container.classList.add( collapsibleOpenClass );
+		container.classList.remove( collapsibleClosedClass );
+
+		toggle.setAttribute( 'aria-expanded', true );
+		panel.setAttribute( 'aria-hidden', false );
+
+		if ( container.classList.contains( 'is-style-preview' ) ) {
+			toggle.innerHTML = toggle.getAttribute("data-close-text");
+		}
+		//dispatch the event on the dom element
+		container.dispatchEvent( eventOpen );
+	};
+
+	/**
+	 * Close Collapsible block
+	 *
+	 * @param object collapsible block
+	 */
+	var closeCollapsible = function( collapsible ) {
+		const container = collapsible.container;
+		const toggle = collapsible.toggle;
+		const panel = collapsible.panel;
+
+		container.classList.remove( collapsibleOpenClass );
+		container.classList.add( collapsibleClosedClass );
+		toggle.setAttribute( 'aria-expanded', false );
+		panel.setAttribute( 'aria-hidden', true );
+
+		if ( container.classList.contains( 'is-style-preview' ) ) {
+			toggle.innerHTML = toggle.getAttribute("data-open-text");
+		}
+		//dispatch the event on the dom element
+		container.dispatchEvent( eventClose );
+	};
+
+	/**
+	 * Toggle collapsible block
+	 *
+	 * @param element collapsible block
+	 */
+	var toggleCollapsible = function( collapsible ) {
+		if ( isOpen( collapsible ) ) {
+			closeCollapsible( collapsible );
+		} else {
+			openCollapsible( collapsible );
+		}
+	};
+
+	/**
+	 * Find all Collapsible blocks
+	 */
+	var findElements = function() {
+		var containers = document.querySelectorAll( '.js-wp-block-bu-collapsible' );
+
+		// Don't coninue if no Collapsible blocks exist
+		if ( containers.length === 0 ) {
+			return;
+		}
+
+		containers.forEach( function( element, i ) {
+			var block = {};
+
+			block.container = element;
+			block.toggle = element.querySelector( '.js-bu-block-collapsible-toggle' );
+			block.panel = element.querySelector( '.js-bu-block-collapsible-content' );
+			collapsibleBlocks.push( block );
+		} );
+	};
+
+	/**
+	 * Set up handlers, aria, and other functionality
+	 */
+	var setupCollapsibleBlocks = function() {
+		if ( collapsibleBlocks.length === 0 ) {
+			return;
+		}
+
+		collapsibleBlocks.forEach( function( collapsible, i ) {
+			const container = collapsible.container;
+			const toggle = collapsible.toggle;
+			const panel = collapsible.panel;
+
+			// Add toggle event
+			toggle.addEventListener( 'click', function( e ) {
+				e.preventDefault();
+				toggleCollapsible( collapsible );
+			} );
+
+			// Set ARIA attributes
+			toggle.setAttribute( 'aria-controls', panel.id );
+			panel.setAttribute( 'aria-labelledby', toggle.id );
+
+			// Setup initial state of each block.
+			if ( isOpenDefault( collapsible ) ) {
+				openCollapsible( collapsible );
+			} else {
+				closeCollapsible( collapsible );
+			}
+
+			if ( isOpen( collapsible ) ) {
+				toggle.setAttribute( 'aria-expanded', true );
+				panel.setAttribute( 'aria-hidden', false );
+			} else {
+				toggle.setAttribute( 'aria-expanded', false );
+				panel.setAttribute( 'aria-hidden', true );
+			}
+		} );
+	};
+
+	/**
+	 * Init
+	 */
+	var collapsibleInit = function() {
+		findElements();
+		setupCollapsibleBlocks();
+	};
+
+	// Start things on dom ready.
+	document.addEventListener( "DOMContentLoaded", function() {
+		collapsibleInit();
+	});
+
+	return {
+		getcollapsibleBlocks: function() {
+			return collapsibleBlocks;
+		},
+		toggleCollapsible: function( collapsible ) {
+			if( collapsible ) {
+				toggleCollapsible( collapsible );
+			}
+		}
+	};
+
+} )();
+;bu_blocks.drawer = ( function() {
 	var drawerBlocks = []; //stores all of our found blocks
 	var $body = document.getElementsByTagName( 'body' )[0]; //target body tag
 	var eventOpen = new CustomEvent( 'bu-blocks-drawer-open' );